Fix token refresh for method permission
